# front/src/api/models/clone-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tickets API
 * REST API для управления билетами, людьми, событиями и площадками.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface CloneRequest
 */
export interface CloneRequest {
  /**
   *
   * @type {number}
   * @memberof CloneRequest
   */
  ticketId: number;
}