# front/src/api/models/ticket.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tickets API
 * REST API для управления билетами, людьми, событиями и площадками.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { Coordinates } from './coordinates';
// May contain unused imports in some cases
// @ts-ignore
import type { Event } from './event';
// May contain unused imports in some cases
// @ts-ignore
import type { Person } from './person';
// May contain unused imports in some cases
// @ts-ignore
import type { Venue } from './venue';

/**
 *
 * @export
 * @interface Ticket
 */
export interface Ticket {
  /**
   *
   * @type {number}
   * @memberof Ticket
   */
  id?: number | null;
  /**
   *
   * @type {string}
   * @memberof Ticket
   */
  name: string;
  /**
   *
   * @type {number}
   * @memberof Ticket
   */
  price: number;
  /**
   *
   * @type {string}
   * @memberof Ticket
   */
  type: TicketTypeEnum;
  /**
   *
   * @type {number}
   * @memberof Ticket
   */
  number: number;
  /**
   *
   * @type {number}
   * @memberof Ticket
   */
  discount?: number | null;
  /**
   *
   * @type {string}
   * @memberof Ticket
   */
  comment?: string | null;
  /**
   *
   * @type {string}
   * @memberof Ticket
   */
  creationDate?: string | null;
  /**
   *
   * @type {Coordinates}
   * @memberof Ticket
   */
  coordinates: Coordinates;
  /**
   *
   * @type {Person}
   * @memberof Ticket
   */
  person?: Person;
  /**
   *
   * @type {Event}
   * @memberof Ticket
   */
  event?: Event;
  /**
   *
   * @type {Venue}
   * @memberof Ticket
   */
  venue?: Venue;
}

export const TicketTypeEnum = {
  VIP: 'VIP',
  USUAL: 'USUAL',
  BUDGETARY: 'BUDGETARY',
  CHEAP: 'CHEAP',
} as const;

export type TicketTypeEnum = (typeof TicketTypeEnum)[keyof typeof TicketTypeEnum];